Fix address truncation

+++ lib/src/jni/idpass_jni.cpp
@@ -58,7 +58,7 @@ jlong idpass_init(JNIEnv *env,
 
     if (ctx) {
         LOGI("idpass_api_init ok");
-        return reinterpret_cast<long>(ctx); // no error
+        return reinterpret_cast<long long>(ctx); // no error
     } else {
         LOGI("idpass_api_init fail: sodium_init");
         return 0;
